Change ordering and normalisation of axes
Axes are now in the same order as those on the Xbox 360
controller. Was seeing some games not work properly with
GC controllers due to axes being wrong.

I have confirmed that these drivers allow Dolphin to work
correctly still, so there should be no downside to this,
just greater compatability.